+++ exporters/handle/src/OpenTelemetry/Exporter/Handle.hs
@@ -1,7 +1,10 @@
-module OpenTelemetry.Exporter.Handle
-  {-# DEPRECATED "use OpenTelemetry.Exporter.Handle.Span instead" #-} (
+module OpenTelemetry.Exporter.Handle (
   module OpenTelemetry.Exporter.Handle.Span,
+  module OpenTelemetry.Exporter.Handle.Metric,
+  module OpenTelemetry.Exporter.Handle.LogRecord,
 ) where
 
+import OpenTelemetry.Exporter.Handle.LogRecord
+import OpenTelemetry.Exporter.Handle.Metric
 import OpenTelemetry.Exporter.Handle.Span